Remove some redundant casts
Explicitly casting  something (variable, expression, ...)  to the same
type as that "something" already is, is fairly pointless.

+++ src/SvgImage.cpp
@@ -83,7 +83,7 @@ namespace tgui
     Vector2f SvgImage::getSize() const
     {
         if (m_data->svg)
-            return {static_cast<float>(m_data->svg->width), static_cast<float>(m_data->svg->height)};
+            return {m_data->svg->width, m_data->svg->height};
         return {0, 0};
     }
 
@@ -101,8 +101,8 @@ namespace tgui
         size.x = static_cast<unsigned int>(size.x * fontScale);
         size.y = static_cast<unsigned int>(size.y * fontScale);
 
-        const float scaleX = size.x / static_cast<float>(m_data->svg->width);
-        const float scaleY = size.y / static_cast<float>(m_data->svg->height);
+        const float scaleX = size.x / m_data->svg->width;
+        const float scaleY = size.y / m_data->svg->height;
 
         auto pixels = MakeUniqueForOverwrite<unsigned char[]>(size.x * size.y * 4);
         nsvgRasterizeXY(m_data->rasterizer, m_data->svg, 0, 0, scaleX, scaleY,
